# games/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Game API",
        "description": "Stephen Game API",
        "contact": {
            "email": "stephengordon48@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/games": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Displays all the games",
                "operationId": "93736f40a135111a310e4b25ad61f30b",
                "responses": {
                    "200": {
                        "description": "Successful operation, Returns a list of Games in JSON format"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Create a new Game",
                "description": "Stores the game in the DB",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "store_id",
                                    "platforms",
                                    "category",
                                    "price"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Example description"
                                    },
                                    "store_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "platforms": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "category": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sports"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/games/{id}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Display the specified resource.",
                "description": "Gets a game by ID",
                "operationId": "f0bce1dcff3d56d19826d671761a43ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Games"
                ],
                "summary": "Create a new Game",
                "description": "Stores the game in the DB",
                "operationId": "put",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "title",
                                    "description",
                                    "category",
                                    "price"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Example description"
                                    },
                                    "category": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sports"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Games"
                ],
                "summary": "Delete a Game",
                "description": "Delete a Game By ID",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a Game",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "204"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}